requirements.txt
dgl==2.4.0+cu121
easydict==1.13
einops==0.8.1
ftfy==6.3.1
gdown==5.2.0
h5py==3.12.1
MinkowskiEngine==0.5.4
numpy==2.2.6
open3d==0.18.0
Pillow==11.2.1
PyYAML==6.0.2
PyYAML==6.0.2
regex==2024.11.6
scikit_learn==1.7.0
scipy==1.15.3
setuptools==72.1.0
timm==1.0.15
torch==2.5.1+cu121
torch.redstone==0.0.6
torchvision==0.20.1+cu121
tqdm==4.67.1
